fin reorganisation parsing requete, semble ok, implementation check erreur de requetes incompletes
